+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.983.4](https://github.com/bigcommerce/checkout-sdk-js/compare/v1.983.3...v1.983.4) (2026-09-24)
+
+### [1.983.3](https://github.com/bigcommerce/checkout-sdk-js/compare/v1.983.2...v1.983.3) (2026-09-23)
+
+
+### Bug Fixes
+
+* **payment:** PAYPAL-7076 send Google Pay billing address on PayPal approve ([29887a3](https://github.com/bigcommerce/checkout-sdk-js/commit/29887a3f86fc8f3b8bfe63697e8222dafc72ec06))
+
+### [1.983.2](https://github.com/bigcommerce/checkout-sdk-js/compare/v1.983.1...v1.983.2) (2026-09-23)
+
+
+### Code Refactoring
+
+* **checkout:** CHECKOUT-10429 Drop Offline Payment Methods for Invoice Flow ([#3419](https://github.com/bigcommerce/checkout-sdk-js/issues/3419)) ([c82e3b4](https://github.com/bigcommerce/checkout-sdk-js/commit/c82e3b47d1b5614f756f79c6cef9addfe329cf30))
+
 ### [1.983.1](https://github.com/bigcommerce/checkout-sdk-js/compare/v1.983.0...v1.983.1) (2026-09-23)